multimodal_sentiment/requirements.txt
transformers==4.37.0
torch==2.2.0
opencv-python==4.9.0.82
librosa==0.10.3
numpy==1.26.4
gradio==4.14.0
deepface==0.0.82
openai-whisper==20231215
moviepy==1.0.3
pandas==2.1.5
